Rename file to match type name
